# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Synops Praxis LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List all partners (super admin only)
      responses:
        "200":
          description: List of partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Partner"
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create a new partner tenant (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /partners/{partnerId}:
    get:
      operationId: getPartner
      tags: [partners]
      summary: Get partner details
      parameters:
        - name: partnerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Partner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
    patch:
      operationId: updatePartner
      tags: [partners]
      summary: Update partner (super admin or partner admin)
      parameters:
        - name: partnerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Partner"
  /partners/{partnerId}/stats:
    get:
      operationId: getPartnerStats
      tags: [partners]
      summary: Partner aggregate stats (de-identified)
      parameters:
        - name: partnerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Aggregated partner stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerStats"

  # ── Organisations ──────────────────────────────────────────────────────────
...
  /studio/generate-script:
    post:
      operationId: generateScript
      tags: [studio]
      summary: Generate script beats from source document text (Claude)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateScriptInput"
      responses:
        "200":
          description: Generated script draft with beats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptDraft"
...
  /studio/scripts/{draftId}:
    get:
      operationId: getScriptDraft
      tags: [studio]
      summary: Get a script draft
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Script draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptDraft"
    patch:
      operationId: updateScriptDraft
      tags: [studio]
      summary: Update/edit a script draft before publishing
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptDraftUpdate"
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptDraft"
  /studio/scripts/{draftId}/publish:
    post:
      operationId: publishScriptDraft
      tags: [studio]
      summary: Convert approved script draft into a module with beats
      parameters:
        - name: draftId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishDraftInput"
      responses:
        "201":
          description: Module created from draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleDetail"

  # ── Socratic Learning Sessions ─────────────────────────────────────────────
...
  /credentials/{credentialId}:
    get:
      operationId: getCredential
      tags: [credentials]
      summary: Get credential details (authenticated)
      parameters:
        - name: credentialId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
  /verify/{credentialId}:
    get:
      operationId: verifyCredential
      tags: [credentials]
      summary: Public credential verification — no auth required
      parameters:
        - name: credentialId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public verification record with evidence summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialVerification"
        "404":
          description: Not found or revoked

  # ── Coach Dashboard ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Partner:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [active, suspended, onboarding]
        contactEmail:
          type: ["string", "null"]
        orgCount:
          type: integer
        learnerCount:
          type: integer
        createdAt:
          type: string
    PartnerInput:
      type: object
      required: [name, slug, contactEmail]
      properties:
        name:
          type: string
        slug:
          type: string
        contactEmail:
          type: string
    PartnerUpdate:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: string
        status:
          type: string
          enum: [active, suspended, onboarding]
    PartnerStats:
      type: object
      required: [partnerId, totalLearners, activeEnrolments, credentialsIssued, completionRate, orgCount]
      properties:
        partnerId:
          type: string
        totalLearners:
          type: integer
        activeEnrolments:
          type: integer
        credentialsIssued:
          type: integer
        completionRate:
          type: number
        orgCount:
          type: integer

    # ── Organisations ──────────────────────────────────────────────────────
...
    ModuleDetail:
      type: object
      required: [id, courseId, title, status, order, beats, createdAt]
      properties:
        id:
          type: string
        courseId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        order:
          type: integer
        beats:
          type: array
          items:
            $ref: "#/components/schemas/Beat"
        estimatedMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string

    # ── Beats ──────────────────────────────────────────────────────────────
    Beat:
      type: object
      required: [id, moduleId, type, order, title, narration]
      properties:
        id:
          type: string
        moduleId:
          type: string
        type:
          type: string
          enum: [title_card, points, scenario, compare, diagram, close]
        order:
          type: integer
        title:
          type: string
        narration:
          type: string
        bulletPoints:
          type: array
          items:
            type: string
        scenario:
          type: ["string", "null"]
        visualData:
          description: Arbitrary JSON for template rendering
        audioUrl:
          type: ["string", "null"]
        audioStatus:
          type: string
          enum: [none, pending, ready, error]
...
    BeatUpdate:
      type: object
      properties:
        title:
          type: string
        narration:
          type: string
        bulletPoints:
          type: array
          items:
            type: string
        scenario:
          type: string
        order:
          type: integer

    # ── Studio ─────────────────────────────────────────────────────────────
    GenerateScriptInput:
      type: object
      required: [sourceText, title]
      properties:
        title:
          type: string
        sourceText:
          type: string
        language:
          type: string
          default: "en-ZA"
        maxBeats:
          type: integer
          default: 6
    ScriptDraft:
      type: object
      required: [id, title, status, beats, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        sourceText:
          type: string
        status:
          type: string
          enum: [generating, ready, published]
        beats:
          type: array
          items:
            $ref: "#/components/schemas/Beat"
        tenantId:
          type: string
        createdAt:
          type: string
    ScriptDraftUpdate:
      type: object
      properties:
        title:
          type: string
        beats:
          type: array
          items:
            $ref: "#/components/schemas/BeatUpdate"
    PublishDraftInput:
      type: object
      required: [courseId]
      properties:
        courseId:
          type: string
        moduleTitle:
          type: string
        # The server has always read lessonType from this body
        # (routes/studio.ts: `const { courseId, moduleTitle, lessonType } = req.body`)
        # but the spec never declared it, so the generated client rejected it and
        # Studio could not choose a lesson type when publishing.
        lessonType:
          type: string
          enum: [socratic, video, slides, quiz]

    # ── Sessions ───────────────────────────────────────────────────────────
...
    Credential:
      type: object
      required: [id, userId, moduleId, issuedAt, decayDate, status, verificationUrl]
      properties:
        id:
          type: string
        userId:
          type: string
        moduleId:
          type: string
        moduleTitle:
          type: string
        issuedAt:
          type: string
        decayDate:
          type: string
        status:
          type: string
          enum: [valid, expired, revoked]
        masteryScore:
          type: number
        evidenceSummary:
          type: string
        verificationUrl:
          type: string
        badgeUrl:
          type: ["string", "null"]
        partnerName:
          type: string
    CredentialVerification:
      type: object
      required: [credentialId, holderName, moduleTitle, issuedAt, decayDate, status, evidenceItems, partnerName]
      properties:
        credentialId:
          type: string
        holderName:
          type: string
        moduleTitle:
          type: string
        issuedAt:
          type: string
        decayDate:
          type: string
        status:
          type: string
          enum: [valid, expired, revoked]
        masteryScore:
          type: number
        evidenceItems:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        partnerName:
          type: string
        verifiedAt:
          type: string
    EvidenceItem:
      type: object
      required: [type, description, recordedAt]
      properties:
        type:
          type: string
          enum: [session_response, assessment_result, coach_attestation, submission_review]
        description:
          type: string
        recordedAt:
          type: string
        score:
          type: ["number", "null"]

    # ── Coach ──────────────────────────────────────────────────────────────
...